refactor: make pipeline and model-spec path linear to read (P06)
- extract compute_binary_metrics to pipeline/metrics.py; evaluate no longer imports from a sibling step
- narrow except Exception to (subprocess.SubprocessError, OSError) in _git_sha(); fix test mock that relied on broad exception swallowing

Closes #124
